refactor: 💡 folder structure
